+++ pkg/apis/replication/v2alpha1/replication_status.go
@@ -36,7 +36,7 @@ type DeploymentReplicationStatus struct {
 	// Destination contains the detailed status of the destination endpoint
 	Destination EndpointStatus `json:"destination"`
 
-	// CancelFailures records the number of times that the configuration was canceled
+	// Deprecated: CancelFailures records the number of times that the configuration was canceled
 	// which resulted in an error.
 	CancelFailures int `json:"cancel-failures,omitempty"`