+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - (Feature) Add HostAliases for Sync
 - (Bugfix) Always stop Sync if disabled
 - (Bugfix) Fix checksum of accepted spec
+- (Feature) Ensure consitency during replication cancellation
 
 ## [1.2.17](https://github.com/arangodb/kube-arangodb/tree/1.2.17) (2022-09-22)
 - (Feature) Add new field to DeploymentReplicationStatus with details on DC2DC sync status=